Add a status icon and reroute all messages to it
Instead of printing out error messages to the console, use
a status icon and reroute the various status messages to the
status icon tooltip.

This will work better with different icons for different
states.

Also bring up the Tweet window when activating the status
icon.

(cherrypick merge - had to move the TwitterClient* to the public
struct, and re-added various signal/class callbacks to tweet-window.c
so that the status icon functionality remained the same as in
mainline.)